Bugfix
The attendance system treats sign outs on another day as if they signed
out the moment they signed in (to give them 0 hours). This tripped up
the attendance parser. This prevents the trip-up by only causing the
attendance parser to sense a new day if the new day starts with a "sign
in"